Enlarge nonce space to 64 bit

+++ src/cryptonote_basic/miner.h
@@ -123,7 +123,7 @@ namespace cryptonote
     epee::critical_section m_template_lock;
     block m_template;
     std::atomic<uint32_t> m_template_no;
-    std::atomic<uint32_t> m_starter_nonce;
+    std::atomic<uint64_t> m_starter_nonce;
     difficulty_type m_diffic;
     uint64_t m_height;
     volatile uint32_t m_thread_index;
